refactor: change data type for force components

+++ src/preprocessor/boundaries/force.cpp
@@ -3,7 +3,7 @@
 #include "error.h"
 #include "force.h"
 
-Force::Force(int i, float x, float y) : index(i), force_x(x), force_y(y) {};
+Force::Force(int i, double x, double y) : index(i), force_x(x), force_y(y) {};
 
 Force Force::createForce(std::unordered_map<std::string, std::optional<double>> force_components) {
 	Force::checkParameters(force_components["index"], force_components["Fx"], force_components["Fy"]);
@@ -15,7 +15,7 @@ Force Force::createForce(std::unordered_map<std::string, std::optional<double>> 
 	);
 }
 
-void Force::checkParameters(std::optional<int> i, std::optional<float> x, std::optional<float> y) {
+void Force::checkParameters(std::optional<int> i, std::optional<double> x, std::optional<double> y) {
 	if (!i.has_value() || !x.has_value() || !y.has_value()) {
 		throw PreprocessorError("Missing required parameter in any force definition");
 	}
@@ -25,11 +25,11 @@ int Force::getIndex() const {
 	return this->index;
 }
 
-float Force::getForceX() const { 
+double Force::getForceX() const { 
 	return this->force_x;
 }
 
-float Force::getForceY() const { 
+double Force::getForceY() const { 
 	return this->force_y;
 }
